機能アップ

+++ main/0-mainloop.js
@@ -57,33 +57,72 @@ while(true) {
 			search: [ /おやすみなさい/ ], run:(test) => {Pod.talkText("はい。おやすみなさい。");}
 		},
 		{
+			search: [ /おつかれさまです/ ], run:(test) => {Pod.talkText("ありがとうございます。");}
+		},
+		{
+			search: [ /すごいですね/ ], run:(test) => {Pod.talkText("ほめても、何も出ませんよ。");}
+		},
+		{
+			search: [ /かわいいですね/ ], run:(test) => {Pod.talkText("ありがとうございます。心に受け取っておきます。");}
+		},
+		{
+			search: [ /じこしょうかい|なまえ/ ], run:(test) => {Pod.talkText("はい。私の名前はポッド。随行支援ユニットです。");}
+		},
+		{
 			search: [ /てんき|きおん|おんど/ ], run:(test) => {
-				Pod.talkText("はい。");
+				Pod.talkText("はい。", true);
 				Pod.node("./11-get-weather.js \"" + julius_result + "\"");
 			}
 		},
 		{
 			search: [ /なんにち|なんようび/ ], run:(test) => {
-				Pod.talkText("はい。");
+				Pod.talkText("はい。", true);
 				Pod.node("./10-get-time.js \"" + julius_result + "\"");
 			}
 		},
 		{
 			search: [ /しゃっとだうん/ ], run:(test) => {
-				Pod.talkText("シャットダウンします。");
-				Pod.run("sudo shutdown -r now");
+				Pod.talkText("シャットダウンしてよろしいですか。");
+				Pod.run("./2-wake-on-voice.sh");
+				const julius_result = File.loadTextFile(env["JULIUS_RESULT"]);
+				File.deleteFile(env["JULIUS_RESULT"]);
+				if(/おねがい/.test(julius_result)) {
+					Pod.talkText("シャットダウンします。");
+					Pod.run("sudo shutdown -P now");
+				}
+				else {
+					Pod.talkText("要求を破棄しました。");
+				}
 			}
 		},
 		{
 			search: [ /りぶーと/ ], run:(test) => {
-				Pod.talkText("リブートします。");
-				Pod.run("sudo reboot");
+				Pod.talkText("リブートしてよろしいですか。");
+				Pod.run("./2-wake-on-voice.sh");
+				const julius_result = File.loadTextFile(env["JULIUS_RESULT"]);
+				File.deleteFile(env["JULIUS_RESULT"]);
+				if(/おねがい/.test(julius_result)) {
+					Pod.talkText("リブートします。");
+					Pod.run("sudo reboot");
+				}
+				else {
+					Pod.talkText("要求を破棄しました。");
+				}
 			}
 		},
 		{
-			search: [ /ぷろせす/ ], run:(test) => {
-				Pod.talkText("プロセスを終了します。");
-				process.exit();
+			search: [ /ぷろせす|ぷろぐらむ/ ], run:(test) => {
+				Pod.talkText("プロセスを終了してよろしいですか。");
+				Pod.run("./2-wake-on-voice.sh");
+				const julius_result = File.loadTextFile(env["JULIUS_RESULT"]);
+				File.deleteFile(env["JULIUS_RESULT"]);
+				if(/おねがい/.test(julius_result)) {
+					Pod.talkText("プロセスを終了します。");
+					process.exit();
+				}
+				else {
+					Pod.talkText("要求を破棄しました。");
+				}
 			}
 		},
 		{
@@ -91,8 +130,9 @@ while(true) {
 				Pod.talkText("どの意味を調べますか。");
 				Pod.run("./3-voice-record.sh");
 				if(File.isExist(env["RECOGNIZE_FILE"])) {
-					Pod.talkText("はい。");
+					Pod.talkText("はい。", true);
 					Pod.run("./4-speech-to-text.sh");
+					Pod.talkText("調査を開始します。", true);
 					const google_result = File.loadTextFile(env["RECOGNIZE_RESULT"]);
 					File.deleteFile(env["RECOGNIZE_RESULT"]);
 					console.log("google " + google_result);
@@ -108,7 +148,7 @@ while(true) {
 				Pod.talkText("はい。なんでしょう。");
 				Pod.run("./3-voice-record.sh");
 				if(File.isExist(env["RECOGNIZE_FILE"])) {
-					Pod.talkText("はい。");
+					Pod.talkText("はい。", true);
 					Pod.run("./4-speech-to-text.sh");
 					const google_result = File.loadTextFile(env["RECOGNIZE_RESULT"]);
 					File.deleteFile(env["RECOGNIZE_RESULT"]);
@@ -120,6 +160,7 @@ while(true) {
 						Pod.node("./11-get-weather.js \"" + google_result + "\"");
 					}
 					else if(/って何|とは|の意味/.test(google_result)) {
+						Pod.talkText("調査を開始します。", true);
 						Pod.node("./12-get-dictionary.js \"" + google_result + "\"");
 					}
 					else {
